+++ changelog.md
@@ -11,6 +11,7 @@ Codion Change Log
 - EntityPopupMenu removed, EntityViewer added.
 - EntityEditPanel.ControlKeys.DISPLAY_ENTITY_VIEWER and DISPLAY_QUERY_INSPECTOR renamed VIEW_ENTITY and INSPECT_QUERY.
 - EntityTablePanel.ControlKeys.DISPLAY_ENTITY_VIEWER, DISPLAY_QUERY_INSPECTOR and DISPLAY_POPUP_MENU renamed VIEW_ENTITY, INSPECT_QUERY and POPUP_MENU.
+- EntityPanel, EntityTablePanel.writeLegacyPreferences() and applyLegacyPreferences() removed.
 
 ## 0.18.66
 ### is.codion.common.reactive